+++ rules/utils/index.js
@@ -1,0 +1,3 @@
+module.exports.isImportDeclaration = node => node.init ?
+  astUtils.isStaticRequire(node.init) :
+  node.parent.type === 'ImportDeclaration';
